requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
adjusttext==1.1.1
    # via AWARE (pyproject.toml)
aiofiles==23.2.1
    # via AWARE (pyproject.toml)
aiogram==2.25.2
    # via
    #   AWARE (pyproject.toml)
    #   aiogram-dialog
aiogram-dialog==1.9.0
    # via AWARE (pyproject.toml)
aiohttp==3.8.6
    # via aiogram
aiomisc==17.5.19
    # via AWARE (pyproject.toml)
aiorwlock==1.4.0
    # via AWARE (pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
aiosmtplib==3.0.1
    # via AWARE (pyproject.toml)
alembic==1.13.1
    # via AWARE (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via asyncclick
asciitree==0.3.3
    # via zarr
astroplan==0.10
    # via
    #   AWARE (pyproject.toml)
    #   ligo-skymap
astropy==6.1.0
    # via
    #   AWARE (pyproject.toml)
    #   astroplan
    #   astropy-healpix
    #   astroquery
    #   cdshealpix
    #   healpy
    #   lalsuite
    #   ligo-skymap
    #   mocpy
    #   pyvo
    #   reproject
astropy-healpix==1.0.3
    # via
    #   AWARE (pyproject.toml)
    #   ligo-skymap
    #   reproject
astropy-iers-data==0.2024.5.27.0.30.8
    # via astropy
astroquery==0.4.7
    # via AWARE (pyproject.toml)
async-timeout==4.0.3
    # via aiohttp
asyncclick==8.1.7.2
    # via AWARE (pyproject.toml)
attrs==23.2.0
    # via aiohttp
authlib==1.3.0
    # via gcn-kafka
babel==2.9.1
    # via aiogram
backports-tarfile==1.1.1
    # via jaraco-context
beautifulsoup4==4.12.3
    # via astroquery
cachetools==4.2.4
    # via aiogram-dialog
cdshealpix==0.6.5
    # via mocpy
certifi==2024.2.2
    # via
    #   aiogram
    #   gcn-kafka
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   AWARE (pyproject.toml)
    #   dask
cloudpickle==3.0.0
    # via
    #   dask
    #   reproject
colorama==0.4.6
    # via AWARE (pyproject.toml)
colorlog==6.8.2
    # via aiomisc
confluent-kafka==2.4.0
    # via
    #   AWARE (pyproject.toml)
    #   gcn-kafka
contourpy==1.2.1
    # via matplotlib
cryptography==42.0.7
    # via
    #   authlib
    #   igwn-auth-utils
    #   ligo-gracedb
    #   pyopenssl
    #   scitokens
    #   secretstorage
cycler==0.12.1
    # via matplotlib
dask[array]==2024.5.1
    # via reproject
diskcache==5.6.3
    # via AWARE (pyproject.toml)
emoji==2.12.1
    # via AWARE (pyproject.toml)
fasteners==0.19
    # via zarr
fonttools==4.52.4
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.5.0
    # via
    #   dask
    #   reproject
future==1.0.0
    # via ligo-gracedb
gcn-kafka==0.3.3
    # via AWARE (pyproject.toml)
greenlet==3.0.3
    # via sqlalchemy
h5py==3.11.0
    # via ligo-skymap
healpy==1.17.1
    # via
    #   AWARE (pyproject.toml)
    #   ligo-skymap
    #   mhealpy
html5lib==1.1
    # via astroquery
idna==3.7
    # via
    #   anyio
    #   requests
    #   yarl
igwn-auth-utils==1.1.0
    # via ligo-gracedb
importlib-metadata==7.1.0
    # via
    #   dask
    #   keyring
jaraco-classes==3.4.0
    # via keyring
jaraco-context==5.3.0
    # via keyring
jaraco-functools==4.0.1
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.4
    # via aiogram-dialog
jplephem==2.22
    # via AWARE (pyproject.toml)
keyring==25.2.1
    # via astroquery
kiwisolver==1.4.5
    # via matplotlib
lalsuite==7.22
    # via ligo-skymap
ligo-gracedb==2.12.0
    # via ligo-skymap
ligo-segments==1.4.0
    # via
    #   lalsuite
    #   ligo-skymap
    #   lscsoft-glue
    #   python-ligo-lw
ligo-skymap==2.0.0
    # via AWARE (pyproject.toml)
locket==1.0.0
    # via partd
logging-journald==0.6.7
    # via aiomisc
lscsoft-glue==4.0.0
    # via
    #   lalsuite
    #   python-ligo-lw
lxml==5.2.2
    # via AWARE (pyproject.toml)
magic-filter==1.0.12
    # via
    #   aiogram
    #   aiogram-dialog
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
matplotlib==3.9.0
    # via
    #   AWARE (pyproject.toml)
    #   adjusttext
    #   lalsuite
    #   ligo-skymap
    #   matplotlib-scalebar
    #   mocpy
matplotlib-scalebar==0.8.1
    # via AWARE (pyproject.toml)
mhealpy==0.3.2
    # via AWARE (pyproject.toml)
mocpy==0.13.1
    # via AWARE (pyproject.toml)
more-itertools==10.2.0
    # via
    #   jaraco-classes
    #   jaraco-functools
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
networkx==3.3
    # via
    #   ligo-skymap
    #   mocpy
numcodecs==0.12.1
    # via zarr
numpy==1.26.4
    # via
    #   AWARE (pyproject.toml)
    #   adjusttext
    #   astroplan
    #   astropy
    #   astropy-healpix
    #   astroquery
    #   contourpy
    #   dask
    #   h5py
    #   healpy
    #   jplephem
    #   lalsuite
    #   ligo-skymap
    #   lscsoft-glue
    #   matplotlib
    #   numcodecs
    #   pandas
    #   ptemcee
    #   pyerfa
    #   python-ligo-lw
    #   reproject
    #   scipy
    #   zarr
orjson==3.10.3
    # via AWARE (pyproject.toml)
packaging==24.0
    # via
    #   astropy
    #   dask
    #   matplotlib
pandas==2.2.2
    # via AWARE (pyproject.toml)
partd==1.4.2
    # via dask
pathvalidate==3.2.0
    # via AWARE (pyproject.toml)
pendulum==3.0.0
    # via AWARE (pyproject.toml)
pillow==10.3.0
    # via
    #   ligo-skymap
    #   matplotlib
ptemcee==1.0.0
    # via ligo-skymap
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via AWARE (pyproject.toml)
pydantic-core==2.18.2
    # via pydantic
pyerfa==2.0.1.4
    # via astropy
pyjwt==2.8.0
    # via scitokens
pyopenssl==24.1.0
    # via lscsoft-glue
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   lalsuite
    #   matplotlib
    #   pandas
    #   pendulum
    #   python-ligo-lw
    #   time-machine
python-dotenv==1.0.1
    # via AWARE (pyproject.toml)
python-ligo-lw==1.8.3
    # via ligo-skymap
pytz==2024.1
    # via
    #   astroplan
    #   babel
    #   ligo-skymap
    #   pandas
pyvo==1.5.2
    # via astroquery
pyyaml==6.0.1
    # via
    #   AWARE (pyproject.toml)
    #   astropy
    #   dask
    #   python-ligo-lw
redis==5.0.4
    # via AWARE (pyproject.toml)
reproject==0.13.1
    # via ligo-skymap
requests==2.32.2
    # via
    #   astroquery
    #   gcn-kafka
    #   igwn-auth-utils
    #   ligo-gracedb
    #   pyvo
    #   scitokens
safe-netrc==1.0.1
    # via igwn-auth-utils
scipy==1.13.1
    # via
    #   AWARE (pyproject.toml)
    #   adjusttext
    #   lalsuite
    #   ligo-skymap
    #   reproject
scitokens==1.8.1
    # via igwn-auth-utils
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   astroplan
    #   html5lib
    #   ligo-segments
    #   lscsoft-glue
    #   python-dateutil
    #   python-ligo-lw
sniffio==1.3.1
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.30
    # via
    #   AWARE (pyproject.toml)
    #   alembic
sqlcipher3==0.5.2
    # via AWARE (pyproject.toml)
time-machine==2.14.1
    # via pendulum
toolz==0.12.1
    # via
    #   dask
    #   partd
tqdm==4.66.4
    # via
    #   ligo-skymap
    #   python-ligo-lw
typing-extensions==4.12.0
    # via
    #   alembic
    #   emoji
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via
    #   pandas
    #   pendulum
urllib3==2.2.1
    # via requests
uvloop==0.19.0
    # via AWARE (pyproject.toml)
webencodings==0.5.1
    # via html5lib
wheel==0.43.0
    # via AWARE (pyproject.toml)
yarl==1.9.4
    # via aiohttp
zarr==2.18.2
    # via reproject
zipp==3.19.0
    # via importlib-metadata
